refactor: move some Settings integrations components to @/settings/integrations
